Added gradle support
