+++ build.gradle
@@ -28,11 +28,11 @@ test {
 }
 
 application {
-    mainClass.set("seedu.duke.Duke")
+    mainClass.set("Chaterpillar")
 }
 
 shadowJar {
-    archiveBaseName = "duke"
+    archiveBaseName = "chaterpillar"
     archiveClassifier = null
     dependsOn("distZip", "distTar")
 }
