+++ data/history.txt
@@ -1,0 +1,2 @@
+D|false|submit screenshot|8/Feb/2024 02:10 PM
+T|true|read book
