+++ src/main/java/chaterpillar/commands/MarkCommand.java
@@ -1,17 +1,17 @@
-package commands;
+package chaterpillar.commands;
 
 import java.io.IOException;
 
-import exceptions.ChaterpillarException;
-import tasks.Task;
-import tasks.TaskList;
-import ui.Ui;
-import storage.Storage;
+import chaterpillar.exceptions.ChaterpillarException;
+import chaterpillar.tasks.Task;
+import chaterpillar.tasks.TaskList;
+import chaterpillar.ui.Ui;
+import chaterpillar.storage.Storage;
 
-public class UnmarkCommand extends Command {
+public class MarkCommand extends Command {
     private final int index;
 
-    public UnmarkCommand(int index) {
+    public MarkCommand(int index) {
         this.index = index;
     }
     public void execute(TaskList tasks, Ui ui, Storage storage) throws ChaterpillarException, IOException {
@@ -20,9 +20,9 @@ public class UnmarkCommand extends Command {
                     "Sorry! That item does not exist in the list.\n" +
                     "You currently have " + tasks.size() + " tasks in the list.");
         } else {
-            ui.echo("Ok, I've marked this task as not done yet:");
-            Task currTask = tasks.get(index - 1);
-            currTask.unmark();
+            ui.echo("Nice! I've marked this task as done:");
+            Task currTask = tasks.get(index-1);
+            currTask.mark();
             ui.echo(currTask.toString());
             storage.saveAllToFile(tasks);
         }
