+++ src/main/java/chaterpillar/exceptions/ChaterpillarException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package chaterpillar.exceptions;
 
 public class ChaterpillarException extends Exception {
     public ChaterpillarException(String message) {
